requirements.txt
# Python dependencies for Agentic AI Travel Planner

# Web Framework
streamlit>=1.31.0

# Data Validation
pydantic>=2.5.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0
